+++ projects/ui/src/lib/components/po-disclaimer-group/po-disclaimer-group.component.spec.ts
@@ -1,4 +1,4 @@
-import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
+import { ComponentFixture, fakeAsync, flush, TestBed, tick } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { configureTestSuite } from './../../util-test/util-expect.spec';
@@ -51,16 +51,8 @@ describe('PoDisclaimerGroupComponent:', () => {
     });
   });
 
-  it('should be created with 3 disclaimers and default removeAll', () => {
-    expect(nativeElement.querySelectorAll('po-disclaimer').length).toBe(3);
-    expect(nativeElement.querySelectorAll('po-tag').length).toBe(1);
-  });
-
-  it('should be created with 3 disclaimers and without removeAll disclaimer', () => {
-    component.hideRemoveAll = true;
-    fixture.detectChanges();
-    expect(nativeElement.querySelectorAll('po-disclaimer').length).toBe(3);
-    expect(nativeElement.querySelector('.po-disclaimer-danger')).toBeFalsy();
+  it('should be created with 3 tags plus tag default removeAll', () => {
+    expect(nativeElement.querySelectorAll('po-tag').length).toBe(4);
   });
 
   it('should hide disclaimer-group if there are no disclaimers', () => {
@@ -82,7 +74,7 @@ describe('PoDisclaimerGroupComponent:', () => {
     component.onCloseAction(disclaimers[0]);
     fixture.detectChanges();
 
-    expect(nativeElement.querySelectorAll('po-disclaimer').length).toBe(2);
+    expect(nativeElement.querySelectorAll('po-tag').length).toBe(2);
   });
 
   it('should remove all disclaimers', () => {
@@ -92,13 +84,241 @@ describe('PoDisclaimerGroupComponent:', () => {
     expect(nativeElement.querySelectorAll('po-disclaimer').length).toBe(0);
   });
 
+  it('should call handleKeyboardNavigationTag after changes', fakeAsync(() => {
+    const changes: any = {
+      disclaimers: ['test']
+    };
+    spyOn(component, 'handleKeyboardNavigationTag');
+
+    component.ngOnChanges(changes);
+    tick();
+
+    expect(component.handleKeyboardNavigationTag).toHaveBeenCalled();
+  }));
+
+  it('onCloseAction: should remove disclaimer and emit current disclaimers', fakeAsync(() => {
+    spyOn(component.change, <any>'emit');
+    spyOn(component, <any>'focusOnNextTag');
+
+    const disclaimerToRemove = { value: 'north', label: 'Region', property: 'region', hideClose: false };
+    const currentDisclaimers = [component.disclaimers[0], component.disclaimers[1]];
+
+    component.onCloseAction(disclaimerToRemove);
+
+    tick();
+
+    expect(component.disclaimers).toEqual(currentDisclaimers);
+    expect(component.change.emit).toHaveBeenCalledWith(component.disclaimers);
+
+    tick(301);
+
+    expect(component.focusOnNextTag).toHaveBeenCalled();
+  }));
+
+  it('onCloseAction: should emit removedDisclaimer and currentDisclaimers in remove action', () => {
+    spyOn(component.remove, <any>'emit');
+
+    const removedDisclaimer = { value: 'north', label: 'Region', property: 'region', hideClose: false };
+    const currentDisclaimers = [component.disclaimers[0], component.disclaimers[1]];
+
+    component.onCloseAction(removedDisclaimer);
+
+    expect(component.disclaimers).toEqual(currentDisclaimers);
+    expect(component.remove.emit).toHaveBeenCalledWith({ currentDisclaimers, removedDisclaimer });
+  });
+
+  it('should handleKeyDown correctly for Space key', () => {
+    const event = new KeyboardEvent('keydown', { code: 'Space' });
+
+    spyOn(event, 'preventDefault');
+    spyOn(event, 'stopPropagation');
+
+    component['handleKeyDown'](event, [], 0);
+
+    expect(event.preventDefault).toHaveBeenCalled();
+    expect(event.stopPropagation).toHaveBeenCalled();
+  });
+
+  it('should handleKeyDown correctly for ArrowLeft key', () => {
+    const event = new KeyboardEvent('keydown', { key: 'ArrowLeft' });
+    spyOn(component as any, 'handleArrowLeft');
+
+    component['handleKeyDown'](event, [], 0);
+
+    expect(component['handleArrowLeft']).toHaveBeenCalled();
+  });
+
+  it('should handleKeyDown correctly for ArrowRight key', () => {
+    const event = new KeyboardEvent('keydown', { key: 'ArrowRight' });
+    spyOn(component as any, 'handleArrowRight');
+
+    component['handleKeyDown'](event, [], 0);
+
+    expect(component['handleArrowRight']).toHaveBeenCalled();
+  });
+
+  it('should call setTabIndex in next tag if index is 1', () => {
+    const tagRemoveElements = [
+      document.createElement('div'),
+      document.createElement('div'),
+      document.createElement('div')
+    ];
+
+    const indexArrow = 1;
+
+    spyOn(component, 'setTabIndex' as any);
+    spyOn(tagRemoveElements[indexArrow + 1], 'focus');
+
+    component['handleArrowRight'](tagRemoveElements, indexArrow);
+
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[indexArrow], -1);
+    expect(tagRemoveElements[indexArrow + 1].focus).toHaveBeenCalled();
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[indexArrow + 1], 0);
+  });
+
+  it('should call setTabIndex in previous tag if index is 1', () => {
+    const tagRemoveElements = [
+      document.createElement('div'),
+      document.createElement('div'),
+      document.createElement('div')
+    ];
+
+    const indexArrow = 1;
+
+    spyOn(component, 'setTabIndex' as any);
+    spyOn(tagRemoveElements[indexArrow - 1], 'focus');
+
+    component['handleArrowLeft'](tagRemoveElements, indexArrow);
+
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[indexArrow], -1);
+    expect(tagRemoveElements[indexArrow - 1].focus).toHaveBeenCalled();
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[indexArrow - 1], 0);
+  });
+
+  it('should focus on the previous element if the tag length is equal to the closed index', () => {
+    const tagRemoveElements = [
+      document.createElement('div'),
+      document.createElement('div'),
+      document.createElement('div')
+    ];
+    const indexClosed = 3;
+    spyOn(tagRemoveElements[indexClosed - 1], 'focus');
+
+    component['focusOnRemoveTag'](tagRemoveElements, indexClosed);
+
+    expect(tagRemoveElements[indexClosed - 1].focus).toHaveBeenCalled();
+  });
+
+  it('should focus on the current element if the tag length is not equal to the closed index', () => {
+    const tagRemoveElements = [
+      document.createElement('div'),
+      document.createElement('div'),
+      document.createElement('div')
+    ];
+    const indexClosed = 1;
+    spyOn(tagRemoveElements[indexClosed], 'focus');
+
+    component['focusOnRemoveTag'](tagRemoveElements, indexClosed);
+
+    expect(tagRemoveElements[indexClosed].focus).toHaveBeenCalled();
+  });
+
+  it('should focus in previous tag if remove next tag ', () => {
+    const tagRemoveElements = [
+      document.createElement('div'),
+      document.createElement('div'),
+      document.createElement('div')
+    ];
+
+    const initialIndex = 3;
+
+    spyOn(component, 'setTabIndex' as any);
+
+    component['initializeTagRemoveElements'](tagRemoveElements, initialIndex);
+
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[initialIndex - 1], 0);
+  });
+
+  it('should add keydown event listeners', () => {
+    const tagRemoveElements = [document.createElement('div')];
+    const initialIndex = 0;
+    const fakeKeyboardEvent = new KeyboardEvent('keydown');
+
+    spyOn(component as any, 'setTabIndex');
+    spyOn(component as any, 'handleKeyDown');
+
+    component['initializeTagRemoveElements'](tagRemoveElements, initialIndex);
+
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[0], 0);
+
+    tagRemoveElements[0].dispatchEvent(fakeKeyboardEvent);
+
+    expect(component['handleKeyDown']).toHaveBeenCalled();
+  });
+
+  it('should add blur event listeners and call setTabIndex', () => {
+    const tagRemoveElements = [document.createElement('div'), document.createElement('div')];
+    const initialIndex = 0;
+
+    spyOn(component as any, 'setTabIndex');
+
+    component['initializeTagRemoveElements'](tagRemoveElements, initialIndex);
+
+    tagRemoveElements[0].focus();
+    tagRemoveElements[0].dispatchEvent(new Event('blur'));
+
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[0], -1);
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[1], 0);
+  });
+
+  it('should set tab index to 0 for the previous element when initialIndex is not 0', () => {
+    const tagRemoveElements = [document.createElement('div'), document.createElement('div')];
+    const initialIndex = 1;
+
+    spyOn(component as any, 'setTabIndex');
+
+    component['initializeTagRemoveElements'](tagRemoveElements, initialIndex);
+
+    expect(component['setTabIndex']).toHaveBeenCalledWith(tagRemoveElements[0], 0);
+  });
+
+  it('focusOnNextTag: should select attribute unselected with index 0', () => {
+    const tagsFake = document.createElement('div');
+    tagsFake.innerHTML = `
+      <div class="po-tag-remove"></div>
+      <div class="po-tag-remove"></div>
+      <div class="po-tag-remove"></div>
+    `;
+
+    spyOn(component, <any>'focusOnRemoveTag');
+
+    component['focusOnNextTag'](0, 'enter');
+
+    expect(component['focusOnRemoveTag']).toHaveBeenCalled();
+  });
+
+  it('focusOnNextTag: should select attribute unselected whitout index', () => {
+    const tagsFake = document.createElement('div');
+    tagsFake.innerHTML = `
+      <div class="po-tag-remove"></div>
+      <div class="po-tag-remove"></div>
+      <div class="po-tag-remove"></div>
+    `;
+
+    spyOn(component, <any>'focusOnRemoveTag');
+
+    component['focusOnNextTag'](null, 'enter');
+
+    expect(component['focusOnRemoveTag']).toHaveBeenCalled();
+  });
+
   describe('Templates:', () => {
     it(`should set tabindex to 0 if have a disclaimer with 'hideClose'.`, () => {
       component.disclaimers = [{ value: 'po', hideClose: false }];
 
       fixture.detectChanges();
 
-      expect(nativeElement.querySelector('.po-disclaimer-remove[tabindex="0"]')).toBeTruthy();
+      expect(nativeElement.querySelector('.po-tag-remove[tabindex="0"]')).toBeTruthy();
     });
 
     it(`shouldn't set tabindex if disclaimer doesn't have 'hideClose'.`, () => {