+++ projects/ui/src/lib/components/po-tabs/po-tab/po-tab.component.ts
@@ -1,6 +1,7 @@
-import { AfterContentInit, Component, ElementRef } from '@angular/core';
+import { AfterContentInit, Component, ElementRef, OnChanges, SimpleChanges } from '@angular/core';
 
 import { PoTabBaseComponent } from './po-tab-base.component';
+import { PoTabsService } from '../po-tabs.service';
 
 /**
  * @docsExtends PoTabBaseComponent
@@ -9,13 +10,19 @@ import { PoTabBaseComponent } from './po-tab-base.component';
   selector: 'po-tab',
   templateUrl: './po-tab.component.html'
 })
-export class PoTabComponent extends PoTabBaseComponent implements AfterContentInit {
-  constructor(private elementRef: ElementRef) {
+export class PoTabComponent extends PoTabBaseComponent implements AfterContentInit, OnChanges {
+  constructor(private elementRef: ElementRef, private tabsService: PoTabsService) {
     super();
   }
 
   ngAfterContentInit() {
     this.setDisplayOnActive();
+  }
+
+  ngOnChanges(): void {
+    setTimeout(() => {
+      this.tabsService.triggerOnChanges();
+    }, 100);
   }
 
   protected setDisplayOnActive() {