+++ projects/ui/src/lib/components/po-tabs/po-tabs.service.ts
@@ -1,0 +1,13 @@
+import { Injectable } from '@angular/core';
+import { Subject } from 'rxjs';
+
+@Injectable()
+export class PoTabsService {
+  private onChangesTriggeredSource = new Subject<void>();
+
+  onChangesTriggered$ = this.onChangesTriggeredSource.asObservable();
+
+  triggerOnChanges() {
+    this.onChangesTriggeredSource.next();
+  }
+}